seção 9:  DTOs E Data annotations

+++ Back/src/ProEventos.API/Controllers/EventosController.cs
@@ -1,12 +1,9 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using ProEventos.Application.Contratos;
-using ProEventos.Domain;
-using ProEventos.Persistence.Contextos;
+using ProEventos.Application.Dtos;
 
 namespace ProEventos.API.Controllers
 {
@@ -28,8 +25,8 @@ namespace ProEventos.API.Controllers
             try
             {
                 var eventos = await _eventoService.GetAllEventosAsync(true);
-                if (eventos == null) return NotFound("Nenhum evento econtrado.");
-
+                if (eventos == null) return NoContent();
+                
                 return Ok(eventos);
             }
             catch (Exception ex)
@@ -45,7 +42,7 @@ namespace ProEventos.API.Controllers
             try
             {
                 var evento = await _eventoService.GetEventoByIdAsync(id, true);
-                if (evento == null) return NotFound("Nenhum evento econtrado.");
+                if (evento == null) return NoContent();
 
                 return Ok(evento);
             }
@@ -62,7 +59,7 @@ namespace ProEventos.API.Controllers
             try
             {
                 var evento = await _eventoService.GetAllEventosByTemaAsync(tema, true);
-                if (evento == null) return NotFound("Nenhum evento econtrado.");
+                if (evento == null) return NoContent();
 
                 return Ok(evento);
             }
@@ -74,12 +71,12 @@ namespace ProEventos.API.Controllers
         }
 
         [HttpPost]
-        public async Task<IActionResult> Post(Evento model)
+        public async Task<IActionResult> Post(EventoDto model)
         {
              try
             {
                 var evento = await _eventoService.AddEventos(model);
-                if (evento == null) return BadRequest("Erro ao tentar adicionar eventos");
+                if (evento == null) return NoContent();
 
                 return Ok(evento);
             }
@@ -91,12 +88,12 @@ namespace ProEventos.API.Controllers
         }
 
         [HttpPut("{id}")]
-        public async Task<IActionResult> Put(int id, Evento model)
+        public async Task<IActionResult> Put(int id, EventoDto model)
         {
              try
             {
                 var evento = await _eventoService.UpdateEvento(id, model);
-                if (evento == null) return BadRequest("Erro ao tentar adicionar eventos");
+                if (evento == null) return NoContent();
 
                 return Ok(evento);
             }
@@ -112,9 +109,13 @@ namespace ProEventos.API.Controllers
         {
              try
             {
+                var evento = await _eventoService.GetEventoByIdAsync(id, true);
+                if (evento == null) return NoContent();
+
                 return await _eventoService.DeleteEvento(id) ?
                      Ok("Deletado") :
-                     BadRequest("Evento não deletado"); 
+                     throw new Exception
+                     ("Ocorreu um problema não específico ao tentar deletar Evento.");
             }
             catch (Exception ex)
             {
